third_party/blink/web_tests/virtual/blur-on-remove/fast/frames/create-iframe-on-blur.html

<!DOCTYPE html>
<body>
<div id=container>
    <input>
</div>
<script src="../../../../resources/js-test.js"></script>
<script>
description("iframes added to a subtree that's being removed should not load");
var container = document.querySelector('#container');
var input = document.querySelector('input');
input.focus();
input.addEventListener('blur', function() {
    iframe = document.createElement('iframe');
    input.parentNode.appendChild(iframe);
});
container.innerHTML = '';
shouldBeNull("iframe.contentWindow");
</script>
</body>
