third_party/blink/web_tests/virtual/blur-on-remove/fast/frames/create-iframe-on-blur-expected.txt

iframes added to a subtree that's being removed should not load

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS iframe.contentWindow is null
PASS successfullyParsed is true

TEST COMPLETE
